run retry directly in queue instead of celery
